Added top_score column to speed up lookups. Worker for top_score

+++ migrations/20170331095601_wilsons_as_function.js
@@ -4,7 +4,7 @@ exports.up = function(knex, Promise) {
       CREATE FUNCTION wilsons(ups int, downs int) RETURNS numeric LANGUAGE plpgsql IMMUTABLE AS $$
       DECLARE
         z numeric;
-        n int;
+        n numeric;
         p numeric;
         l numeric;
         r numeric;
